feat: include interfaces, move 'functions', 'types', 'interfaces' ts files to helper folder
