+++ news-JS/src/components/helper/functions.ts
@@ -1,0 +1,35 @@
+import Sources from '../view/sources/sources';
+import { sourceType } from './types';
+
+export function sortArray(array: sourceType[], letter: string): sourceType[] {
+    const res: sourceType[] = [];
+    for (let i = 0; i < array.length; i++) {
+        if (array[i].name[0] == letter) {
+            res.push(array[i]);
+        }
+    }
+    return res;
+}
+
+export function showByAlph(sources: sourceType[]) {
+    const alphBlock = document.querySelector('.alph-btns') as HTMLDivElement;
+    const buttonsBlock = document.querySelector('.buttons') as HTMLDivElement;
+    const alph = 'ABCDEFGHILMNPRSTUVWXY' as string
+    const drawClass = new Sources();
+
+    alphBlock.innerHTML = '';
+    for (let i = 0; i < alph.length; i++) {
+        const letter: HTMLButtonElement = document.createElement('button');
+        alphBlock.append(letter);
+        letter.innerHTML = alph[i];
+        letter.className = 'letter';
+    }
+    
+    const allLetters = document.querySelectorAll('.letter') as NodeListOf<Element>;
+    allLetters.forEach((e) => {
+        e.addEventListener('click', () => {
+            buttonsBlock.innerHTML = '';
+            drawClass.draw(sortArray(sources, e.innerHTML));
+        });
+    });
+}
